Add files via upload
Exploratory Data Analysis(EDA) is an approach to analyze the data using visual techniques. It is used to discover trends, patterns, or to check assumptions with the help of statistical summary and graphical representations. By using python libraries, I performed data cleaning, processing, visualization on MYNTRA dataset. From this project, I used to learn how to deal with data.